requirement.txt
streamlit>=1.35.0
tensorflow>=2.12.0
numpy>=1.23.0
Pillow>=9.5.0
av
streamlit-webrtc
opencv-python
